OT231-57: remove Image.class
Now Slide implements IImage
