+++ src/main/java/com/alkemy/ong/domain/Slide.java
@@ -1,5 +1,9 @@
 package com.alkemy.ong.domain;
 
+import java.io.ByteArrayInputStream;
+import java.io.InputStream;
+import java.util.Base64;
+import java.util.UUID;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
@@ -7,11 +11,28 @@ import lombok.Setter;
 @Builder
 @Getter
 @Setter
-public class Slide {
+public class Slide implements IImage{
 
   private String imageUrl;
   private String text;
   private Integer order;
   private String base64File;
+  
+  @Override
+  public InputStream getFile() {
+    byte[] decodedBytes = Base64.getDecoder().decode(base64File);
+    return new ByteArrayInputStream(decodedBytes);
+  }
+
+  @Override
+  public String getFileName() {
+    UUID uuid = UUID.randomUUID();
+    return uuid.toString();
+  }
+
+  @Override
+  public String getContentType() {
+    return "image/jpeg";
+  }
 
 }
